隐藏已评价约会

+++ src/main/java/com/corgi/entity/LikedActivity.java
@@ -7,6 +7,6 @@ public class LikedActivity {
     private String picUrl;
     private String activityId;
     private String category;
-    private Integer height = 0;
-    private Integer width = 0;
+    private Long height = 0L;
+    private Long width = 0L;
 }
